✨ better exception handling
add NMI and HardFault handler, which lead to a bpkt
exception. this should bring up the debugger or (if
not attached) but the core in lockout state, which
often results in a reset. should be fine for now.

add a bpkt instruction to the panic handler, for the
same reason.

add correct symbol types and sizes to the assembly.